Display estimated Active cases; more below
The display now is tabular and shows a lot more of the intermediate
calculation columns than the final product should.

Also, adds the initial code for boilerplate ArgumentParser setup.
This allows the user to specify csv file, and, optionally, the days of
delay from becoming Confirmed to Death or Recovery (two independent
variables with defaults, but may be user specified).