Update README; cleanup $TOP per below
Previous README was A) notes on the data source change that caused the
project to migrate from hand entry of Orange County data to use of the
Hopkins data, and B) notes on the conversion to Python 3.8 and pipenv.

The current README is more in line with GitHub norms.  With the
original work moved under the misc directory in files named to
correspond to A and B in the first paragraph.

Some early notes during the "hand entry of Orange County data phase"
with respect to hospital and testing metrics have also been moved to
the misc directory.

In $TOP, added text for GitHub About.